src/spot_bots/sol_bot_15m/requirements.txt
# Dependencias principales
pandas==1.5.3
numpy==1.23.5
python-binance==1.0.16
python-telegram-bot==13.15
python-dotenv==1.0.0
requests==2.28.2
matplotlib==3.7.1
textblob==0.15.3
tweepy==4.10.0
scikit-learn==1.2.2

# Dependencias para aprendizaje profundo
tensorflow==2.9.1
ccxt==3.0.27
ta-lib==0.4.24
retrying==1.3.4
pytz==2023.3
psutil==5.9.5
joblib==1.2.0

# Dependencias para Google Cloud
google-cloud-storage==2.8.0
google-cloud-logging==3.5.0
google-cloud-secret-manager==2.16.1
google-cloud-monitoring==2.14.2

# Dependencias para manejo de errores y robustez
tenacity==8.2.2
backoff==2.2.1
fastjsonschema==2.16.3

# Dependencias para monitoreo y diagnóstico
prometheus-client==0.16.0
opencensus==0.11.0
opencensus-ext-stackdriver==0.8.0

# Dependencias para optimización
numba==0.56.4
dask==2023.3.2

# Dependencias para serialización avanzada
cloudpickle==2.2.1
msgpack==1.0.5
